requirements.txt
requests>=2.31.0
python-dotenv>=1.1.1
python-vlc>=3.0.241
yt-dlp>=2025.9.1
google-api-python-client>=2.99.0
